Added Debug and "fixed" Pathing destination issue
I fixed the destination issue by setting it thus that in the situations
where the destination and the final point in the path don't match up,
just set them to be equal. I also made it so it checks for this inequal
error and upon finding it, tries again with an adjustment of 0.01 to the
x value of the target destination, which seems to fix the massive error
problem. I suspect it may be an engine bug of some form but until I get
the chance to also test with GDScript, hard to tell.

I also found out how to add and use the debug system from within VSCode
for Godot, so I'm now able to use actual breakpoints and such.

Next I really need to work on the basic AI, I'm going to detail my
current plan in a plan file of some form and then crash for tonight.