Develop random wander method or task
I've got a random wandering method which moves a pawn randomly around a
map. It alongside the "rest" method are going to be the 2 initial tasks
within the finite state machine/GOAP AI I'm building. Once I've got that
I've essentially got the game and then all that's needed is some
emergent interactions.