Implemented a needs system
Implemented a needs system which allows has different needs that tick
down or up at different rates depending on the state (or action) the
pawn currently is in or performing. To complete the A.I all that's left
to do is write a basic evaluation script and a second state for it to
use.

Also stupid C# and it's inconsistencies are driving me mad. I'm ok with
(though not fond of) mutability as it's idiomatic style, ala C#. I'm
also cool with immutability as idiomatic, ala F#. What I'm not cool with
is a language that prohibits use of it's idiomatic tooling cause it's
surprising and frustrating, ala C#. In general you want to mutate in
place for efficiency and such but because a dictionary doesn't let you
customise how you iterate through it (only foreach) you can't mutate it
in place. So you need to copy out part of it to a normal list which you
can then iterate through to allow you to mutate the dictionary!