Create framework for GOAP AI
Created a framework for a GOAP ai, with two states, walkaround and rest.
However there are numerous bugs and the code is rather messy meaning I'm
struggling to solve these issues. Notably I can't get the AI to properly
switch between tasks as their evaluation of needs seems to have some
problems. In addition the language is rather confusing.