Create initial change tile task
I've started 2 primary goals in this branch. First to enable pawns to
modify the tilemap and second, to evaluate the "best" way of having a
dynamic and/or interactive tilemap. How much logic can/should I do with
the tile map and how much should I do with discrete nodes.