Create tilemap scene and specific tile instancing
The game can now have growing trees! And I may have just discovered
interfaces which while a bit of a bother, should be a significant help
with the problems I've been facing still, though it's probably too late
to really help my cause.

To explain I've come to the conclusion that emergent gameplay systems
are just more work to make with a restricted/nominal static typed
language than with either a dynamic typed language or a more advanced
type system. This is because while logically you just want to operate on
what something can do, in a nominal typing system you can only operate
on what something is and what it is doesn't necessarily derive from what
it can do.