Added Export Profile
I narrowed down a certain amount of task related bugs but sadly had to
come to the conclusion that the only way it'd work is if I massively
redid the entire thing. Now if I'd known that you needed to use tasks to
properly stall/store methods, I would have focused on them from the
beginning but I didn't so whoooooooops. I'm not sure that the fact that
the refactoring is so blinking punishing is a good thing or not.