Attempt to implement signal calls in F#
My original design tried to use signals for a lot, as a good way of
communicating between game entities. However I seem to be having
significant trouble doing this in pure F#. As such I'm now going to
pivot to simply trying to update/upgrade my original game with F# code
and see how that goes.