Allow need records to updated from C# calls
There is successful evidence of using abstract member as a shim which is
called from C# land to cause a F# record to be updated with minimal
mutation.

Currently I have a pure/immutable F# record which is wrapped in a
mutable reference. So ideally if I can discover a non mutation
alternative I can just unwind that wrapping.