Created F# functions to adjust growth of needs.
This includes changing the state of decaying, and rate of changing.

I also made a "dummy" bounce function to test this functionality, which
works. I've actually had thoughts about completely overhauling the need
system to work just in terms of magnitudes and then providing a "shim"
which would tell a UI system to either display it as moving a certain
direction.

I'd need to overhaul it all to just work in terms of vectors/magnitudes.
ergo, don't consider the direction, except to provide it as a piece of
information to a UI system. Have 0 as empty, 100 as full. Decaying
always makes things smaller, healing/growing always makes things bigger.

Then in the UI layer, if it's a decay is going up need, like boredom,
simply display it as max-current. Otherwise, just display it as current.