Expand NeedSystem and Start on TaskSystem
I've got most of the need functions written, now I'm just working on
making the tasks to relate to and operate off them.

I'm not certain but I think I might look into make a queue or such for
tasks and then just operate on/through that list to change/use tasks.

I need to update the getPriorityNeed function to make sure it returns a
NeedName, not a float (or maybe a tuple of the two?)

Once I have that, all I need to do is then make the changeTask function
and I should have all the essential building blocks. Then comes the
challenge or task of seeing what else I need to build/shim to use this
to replace the existing C# system.