Working F# task system.
It only has the wonder task currently but it works! Shouldn't be too
hard to add the few extra tasks like naping and cutting trees and we're
good to go. There's a lot of refactoring that could be done but less
than the equivelent point in the C# dev (which pedantically speaking,
was never reached, this works better than the C# code).